DPDK: Upgrade to 20.11.9(LTS).

+++ dpdk/drivers/net/tap/tap_tcmsgs.h
@@ -24,14 +24,14 @@
 
 #define MULTIQ_MAJOR_HANDLE (1 << 16)
 
-void tc_init_msg(struct nlmsg *msg, uint16_t ifindex, uint16_t type,
+void tc_init_msg(struct nlmsg *msg, unsigned int ifindex, uint16_t type,
 		 uint16_t flags);
-int qdisc_list(int nlsk_fd, uint16_t ifindex);
-int qdisc_flush(int nlsk_fd, uint16_t ifindex);
-int qdisc_create_ingress(int nlsk_fd, uint16_t ifindex);
-int qdisc_create_multiq(int nlsk_fd, uint16_t ifindex);
-int qdisc_add_ingress(int nlsk_fd, uint16_t ifindex);
-int qdisc_add_multiq(int nlsk_fd, uint16_t ifindex);
-int filter_list_ingress(int nlsk_fd, uint16_t ifindex);
+int qdisc_list(int nlsk_fd, unsigned int ifindex);
+int qdisc_flush(int nlsk_fd, unsigned int ifindex);
+int qdisc_create_ingress(int nlsk_fd, unsigned int ifindex);
+int qdisc_create_multiq(int nlsk_fd, unsigned int ifindex);
+int qdisc_add_ingress(int nlsk_fd, unsigned int ifindex);
+int qdisc_add_multiq(int nlsk_fd, unsigned int ifindex);
+int filter_list_ingress(int nlsk_fd, unsigned int ifindex);
 
 #endif /* _TAP_TCMSGS_H_ */
